Moving file to archive folder
changes have been made to the load_data.py file as the table structure of state wise data in the 'https://www.mohfw.gov.in/' website has changed. Latest code to web scrape the table from the website has been uploaded in the master branch and old code is moved to archive folder.